+++ api/src/routes/user.js
@@ -11,7 +11,6 @@ const osmesa = require('../services/osmesa')
 const { canEditUser } = require('../passport')
 const db = require('../db/connection')
 const refreshStatus = require('../utils/osmesaStatus.js')
-const getSumEdits = require('../utils/sum_edits')
 const { prop } = require('ramda')
 
 /**
@@ -19,7 +18,7 @@ const { prop } = require('ramda')
  * /user/:id
  *
  * The user stats route will display the OSM stats for a user with id :id. The
- * API will get the stats from OSMESA_API/users/{id}
+ * API will get the stats from the OSMESA database
  *
  * @param {Object} req - the request object
  * @param {Object} res - the response object
@@ -96,7 +95,6 @@ async function get (req, res) {
   }
 
   osmesaData.extent_uri = join(APP_URL_FINAL, `/scoreboard/api/extents/user/${id}/{z}/{x}/{y}.mvt`)
-  osmesaData['edit_sum'] = getSumEdits(osmesaData)
   const refreshDate = await refreshStatus('user')
 
   let badges