+++ api/tests/api/badges.test.js
@@ -13,9 +13,9 @@ const migrationsDirectory = path.join(dbDirectory, 'migrations')
 const seedsDirectory = path.join(dbDirectory, 'seeds', 'test')
 
 test.before(async () => {
-  app = await app()
   await db.migrate.latest({ directory: migrationsDirectory })
   await db.seed.run({ directory: seedsDirectory })
+  app = await app()
   adminUser = await createAuthenticatedUser(app, ['admin'])
 })