+++ components/admin/AdminSectionList.js
@@ -93,6 +93,18 @@ export default function AdminSectionList (props) {
           </li>
         </ul>
       </li>
+      <li className='widget'>
+        <h3 className='header--medium'>
+          Settings
+        </h3>
+        <ul>
+          <li>
+            <Link href='/admin/settings'>
+              <a className='link--normal'>Edit App Settings</a>
+            </Link>
+          </li>
+        </ul>
+      </li>
     </ul>
   )
 }