+++ pages/dashboard.js
@@ -58,9 +58,14 @@ class Dashboard extends Component {
 
   render () {
     if (this.state.loading) {
-      return <LoadingState />
+      return (
+        <div>
+          <header className='header--internal--green header--page' style={{ paddingBottom: '8rem' }} />
+          <LoadingState />
+        </div>
+      )
     }
-    const { authenticatedUser } = this.props
+    const { authenticatedUser, project } = this.props
     const { loggedIn, account } = authenticatedUser
     const { assignments, favorites, country, allCampaigns } = account
 
@@ -90,7 +95,6 @@ class Dashboard extends Component {
       ],
       osmesaData
     )
-
     // Calculate counts for panel
     const badgeCount =
       account.badges && badges.earnedBadges
@@ -105,7 +109,6 @@ class Dashboard extends Component {
         <NotLoggedIn message='Log in with your OSM account to see your personalized dashboard' />
       )
     }
-
     // Dashboard header variables
     let profileImage = null
     let name = null
@@ -127,6 +130,8 @@ class Dashboard extends Component {
       accountId = authenticatedUser.account.id
     }
     const recordsExport = [{ ...account.records, badgeCount, campaignCount, name }]
+    const userHasEdits = edit_times.length > 0
+    const userHasCampaigns = (favorites.length > 0 || assignments.length > 0 || campaignCount > 0)
     return (
       <div className='dashboard'>
         <DashboardHeader
@@ -143,61 +148,67 @@ class Dashboard extends Component {
           facets={[
             { label: 'Campaigns', value: formatDecimal(campaignCount) },
             { label: 'Badges', value: formatDecimal(badgeCount) },
-            { label: 'Edits', value: formatDecimal(osmesaData.edit_sum) },
+            { label: 'Edits', value: formatDecimal(osmesaData.edit_count) },
             { label: 'Changesets', value: formatDecimal(changesetCount) }
           ]}
         />
-        <div className='row'>
-          <DashboardBlurb {...osmesaData} />
+        <div id='blurb' className='row'>
+          <DashboardBlurb {...osmesaData} project={project} />
           <div className='widget-33 page-actions'>
-            <CSVExport filename={`${name}_ScoreboardData.csv`} data={recordsExport} />
+            {userHasEdits &&
+              <CSVExport filename={`${name}_ScoreboardData.csv`} data={recordsExport} />
+            }
           </div>
         </div>
-        <section className='section--dark'>
-          <div className='row'>
-            {isAdmin(authenticatedUser.account.roles) && this.renderAdmin()}
-            <DashboardAssignments
-              favorites={favorites}
-              assignments={assignments}
-              authenticatedUser={authenticatedUser}
-              all={allCampaigns}
-            />
-          </div>
-        </section>
-        <section>
-          <div className='row'>
-            <div className='map-lg'>
-              <UserExtentMap uid={uid} extent={extent_uri} />
+        {
+          (userHasCampaigns || isAdmin(authenticatedUser.account.roles)) &&
+          <section id='admin-assignments' className='section--dark'>
+            <div className='row'>
+              {isAdmin(authenticatedUser.account.roles) && this.renderAdmin()}
+              {userHasCampaigns && this.renderAssignments()}
             </div>
-          </div>
-        </section>
-        <section>
-          <div className='row'>
-            <div className='widget-container'>
-              <div className='widget-66'>
-                <CampaignsChart
-                  campaigns={allCampaigns}
-                  hashtags={hashtags}
-                  height='260px'
-                />
-              </div>
-              <div className='widget-33'>
-                <EditBreakdownChart {...breakdownChartProps} height='260px' />
-              </div>
-            </div>
-          </div>
-        </section>
-        <section>
-          <div className='row widget-container'>
-            <DashboardSidebar teams={teams} osmesaData={osmesaData} />
-            <div className='widget-75'>
-              <DashboardBadges badges={badges} name={name} />
-            </div>
-          </div>
-          <div className='row'>
-            <CalendarHeatmap times={edit_times} />
-          </div>
-        </section>
+          </section>
+        }
+        {
+          userHasEdits
+            ? <>
+              <section id='dashboard_map'>
+                <div className='row'>
+                  <div className='map-lg'>
+                    <UserExtentMap uid={uid} extent={extent_uri} />
+                  </div>
+                </div>
+              </section>
+              <section id='dashboard_charts'>
+                <div className='row'>
+                  <div className='widget-container'>
+                    <div className='widget-66'>
+                      <CampaignsChart
+                        campaigns={allCampaigns}
+                        hashtags={hashtags}
+                        height='260px'
+                      />
+                    </div>
+                    <div className='widget-33'>
+                      <EditBreakdownChart {...breakdownChartProps} height='260px' />
+                    </div>
+                  </div>
+                </div>
+              </section>
+              <section id='dashboard_sidebar-badges-heatmap'>
+                <div className='row widget-container'>
+                  <DashboardSidebar teams={teams} osmesaData={osmesaData} loggedIn={loggedIn} />
+                  <div className='widget-75'>
+                    <DashboardBadges badges={badges} name={name} />
+                  </div>
+                </div>
+                <div className='row'>
+                  <CalendarHeatmap times={edit_times} />
+                </div>
+              </section>
+              </>
+            : this.renderBlankState()
+        }
       </div>
     )
   }
@@ -215,6 +226,27 @@ class Dashboard extends Component {
         <br />
         <br />
       </div>
+    )
+  }
+
+  renderAssignments () {
+    const { authenticatedUser } = this.props
+    const { assignments, favorites, allCampaigns } = authenticatedUser.account
+    return (
+      <DashboardAssignments
+        favorites={favorites}
+        assignments={assignments}
+        authenticatedUser={authenticatedUser}
+        all={allCampaigns}
+      />
+    )
+  }
+
+  renderBlankState () {
+    return (
+      <section className='row'>
+        <img className='img--blank-state' src='static/dashboard-temp/open_maps.svg' />
+      </section>
     )
   }
 }