+++ pages/home-component.js
@@ -10,6 +10,7 @@ import { formatDecimal } from '../lib/utils/format'
 import TopEditorsChart from '../components/charts/TopEditorsChart'
 import EditsByCountry from '../components/charts/EditsByCountryChart'
 import ScoreboardPanel from '../components/ScoreboardPanel'
+import { LoadingState } from '../components/common/LoadingState'
 
 const Map = dynamic(() => import('../components/charts/LeafletHomeMap'), {
   ssr: false
@@ -17,14 +18,46 @@ const Map = dynamic(() => import('../components/charts/LeafletHomeMap'), {
 
 const projectName = process.env.PROJECT_NAME || 'OpenStreetMap'
 export class Home extends Component {
+  constructor () {
+    super()
+    this.state = {
+      loading: true
+    }
+  }
+
   componentDidMount () {
     this.props.getTopStats()
+  }
+
+  componentDidUpdate () {
+    if (this.state.loading && (this.props.topStats)) {
+      this.setState({ loading: false })
+    }
   }
 
   render () {
     const { topStats } = this.props
 
-    if (!topStats) return <div />
+    if (this.state.loading) {
+      return (
+        <div className='home'>
+          <header className='header--homepage header--page'>
+            <div className='row'>
+              <div className='width--shortened'>
+                <div className='section-sub--left'>
+                  <h1 className='header--xxlarge header--with-description'>Tracking Map Edits Around the World</h1>
+                  <p className='description--header'>{'See what’s happening throughout the ' + projectName + ' ecosystem. From which campaigns are the most active, to detailed information about the contributing mappers.'}</p>
+                  <Link href='/about'>
+                    <a className='link--large'>Learn More</a>
+                  </Link>
+                </div>
+              </div>
+            </div>
+          </header>
+          <LoadingState />
+        </div>
+      )
+    }
 
     const { numCampaigns, priorityCampaigns, numUsers, features, topEdits, editsByCountry, totalEdits, numCountries } = topStats
 
@@ -53,7 +86,7 @@ export class Home extends Component {
           <div className='row'>
             <div className='width--shortened'>
               <h2 className='header--large'>Campaigns</h2>
-              <div className='home-map' style={{ width: '100%', height: '375px', 'marginBottom': '40px' }}>
+              <div className='home-map'>
                 {features
                   ? <Map overlay={features} />
                   : <div>Loading map...</div>
@@ -100,14 +133,14 @@ export class Home extends Component {
         </section>
         <section>
           <div className='row'>
-            <div className='width--shortened graphs--users widget widget-container'>
+            <div className='width--shortened widget widget-container'>
               <h2 className='header--large widget-100'>Leaderboard</h2>
               <div className='widget-25'>
-                <h3>Countries Mapped</h3>
+                <h3 className='header--small'>Countries Mapped</h3>
                 <EditsByCountry edits={editsByCountry} />
               </div>
               <div className='widget-75 chart' style={{ height: '430px', marginBottom: '50px' }}>
-                <h3>Editors</h3>
+                <h3 className='header--small'>Editors</h3>
                 {
                   topEdits ? <TopEditorsChart edits={topEdits} /> : <div>Loading...</div>
                 }